+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+  - Edge page - Allow creation of Nuvlabox without VPN
   - Notifications page - open first accordion, added refresh buttons,
     displaying number of actual subscriptions per configuration.